Revert 46 revert 42 sofia dev2

+++ attack_run.py
@@ -80,7 +80,20 @@ def main(cfg: DictConfig):
     )
 
     criterion = get_criterion(cfg["criterion_name"], cfg["criterion_params"])
-    
+    if cfg['pretrained_disc']:
+        model_start_name = 'model_{}_{}_{}_attack_{}'.format(
+            cfg["attack_model"]["name"],
+            cfg["model_id_attack"],
+            cfg["dataset"]["name"],
+            'fgsm_attack',
+        )
+        model_add_name = cfg['disc_parametr']
+        task_name = model_start_name + model_add_name
+        path = weights_from_clearml_by_name(cfg['project_weights_disc'],task_name, 'disc_weights')
+        
+    else:
+        path = cfg["disc_path"]
+
     if cfg["use_disc_check"]:
         disc_check_list = get_disc_list(
             model_name=cfg["disc_model_check"]["name"],
@@ -89,6 +102,7 @@ def main(cfg: DictConfig):
             device=device,
             path=cfg["disc_path"],
             train_mode=False,
+            from_clearml=False
         )
     else:
         disc_check_list = None
@@ -112,8 +126,9 @@ def main(cfg: DictConfig):
             model_params=cfg["disc_model_reg"]["params"],
             list_disc_params=cfg["attack"]["list_reg_model_params"],
             device=device,
-            path=cfg["disc_path"],
+            path=path,
             train_mode=cfg["disc_model_reg"]["attack_train_mode"],
+            from_clearml=False
         )
 
     attack = get_attack(cfg["attack"]["name"], attack_params)
@@ -159,7 +174,10 @@ def main(cfg: DictConfig):
         for param in cfg['attack']['named_params']:
             attack_metrics[f'{param}_param'] = round(cfg['attack']['attack_params'][param], 4)
         save_attack_metrics(attack_metrics, cfg["save_path"], attack_save_name)
-    os.remove(attack_model_path)
+    if cfg['pretrained']:
+        os.remove(attack_model_path)
+    if cfg['pretrained_disc']:
+        os.remove(path)
 
 if __name__ == "__main__":
     main()
